refactor: 단위테스트 클래스 분리
